all Done

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -5,7 +5,7 @@
  */
 void puts_half(char *str)
 {
-	int len = 0, i;
+	long int len = 0, i;
 
 	while (str[len] != '\0')
 	{
